refactor: useState は Quagga には不要だった
